Properize the filenames as they are sideloaded
Many of the images we deal with have spaces in them. In order for the sideload to work properly, we need to replace those spaces with %20.

Once the spaces have %20 in them, they get confusing because they are saved to the filesystem that way and then the server is actually thinking it should look for a file with a space in it rather than a %20.

By sanitizing the filename to replace the %20 with a -, we have something the server and the browser appreciate.

We'll still need to track this as we go so that we can replace with a proper URL at some point.